More safely write files & overwrite broken JSON
Now, use tempfile to write JSON to a "safe" file first; and only if
that fully succeeds, move it to its final resting place. In addition,
parse any JSON we already have and overwrite it if it doesn't parse
right - whatever we have is likely to be better than what's on disk.